+++ multiple-document-ownership-config.js
@@ -13,30 +13,17 @@ export const WithOwnershipConfig = (ConfigClass) => {
 
         getData(options) {
             // User permission levels
-            let playerLevels = {};
-            if (isNewerVersion(game.version, "9.999999")) {
-                playerLevels = Object.entries(CONST.DOCUMENT_META_OWNERSHIP_LEVELS).map(([name, level]) => {
-                    return { level, label: game.i18n.localize(`OWNERSHIP.${name}`) };
-                });
-                for (let [name, level] of Object.entries(CONST.DOCUMENT_OWNERSHIP_LEVELS)) {
-                    if ((level < 0)) continue;
-                    playerLevels.push({ level, label: game.i18n.localize(`OWNERSHIP.${name}`) });
-                }
-            } else {
-                playerLevels["-2"] = game.i18n.localize("PERMISSION.DEFAULT");
-                playerLevels["-1"] = game.i18n.localize("PERMISSION.NOCHANGE");
-                for (let [n, l] of Object.entries(CONST.DOCUMENT_PERMISSION_LEVELS)) {
-                    playerLevels[l] = game.i18n.localize(`PERMISSION.${n}`);
-                }
+            let playerLevels = Object.entries(CONST.DOCUMENT_META_OWNERSHIP_LEVELS).map(([name, level]) => {
+                return { level, label: game.i18n.localize(`OWNERSHIP.${name}`) };
+            });
+            for (let [name, level] of Object.entries(CONST.DOCUMENT_OWNERSHIP_LEVELS)) {
+                if ((level < 0)) continue;
+                playerLevels.push({ level, label: game.i18n.localize(`OWNERSHIP.${name}`) });
             }
 
             // Default permission levels
             const defaultLevels = foundry.utils.deepClone(playerLevels);
-            if (isNewerVersion(game.version, "9.999999")) {
-                defaultLevels.shift();
-            } else {
-                delete defaultLevels["-2"];
-            }
+            defaultLevels.shift();
 
             // Player users
             const users = game.users.map(user => {
@@ -50,7 +37,7 @@ export const WithOwnershipConfig = (ConfigClass) => {
             // Construct and return the data object
             return {
                 currentDefault: CONST.DOCUMENT_META_OWNERSHIP_LEVELS?.NOCHANGE ?? "-1",
-                instructions: game.i18n.localize(isNewerVersion(game.version, "9.999999") ? "OWNERSHIP.HintDocument" : "PERMISSION.HintDocument"),
+                instructions: game.i18n.localize("OWNERSHIP.HintDocument"),
                 defaultLevels,
                 playerLevels,
                 isFolder: false,
@@ -62,7 +49,7 @@ export const WithOwnershipConfig = (ConfigClass) => {
             event.preventDefault();
             if (!game.user.isGM) throw new Error("You do not have the ability to configure permissions.");
 
-            const metaLevels = isNewerVersion(game.version, "9.999999") ? CONST.DOCUMENT_META_OWNERSHIP_LEVELS : { DEFAULT: -2, NOCHANGE: -1 };
+            const metaLevels = CONST.DOCUMENT_META_OWNERSHIP_LEVELS;
             const omit = metaLevels.NOCHANGE;
             const ownershipLevels = {};
             for (let [user, level] of Object.entries(formData)) {
@@ -75,16 +62,12 @@ export const WithOwnershipConfig = (ConfigClass) => {
 
             let cls = getDocumentClass(this.documents[0].documentName);
             const updates = this.documents.map(d => {
-                const ownership = foundry.utils.deepClone(isNewerVersion(game.version, "9.999999") ? d.ownership : d.data.permission);
+                const ownership = foundry.utils.deepClone(d.ownership);
                 for (let [k, v] of Object.entries(ownershipLevels)) {
                     if (v === metaLevels.DEFAULT) delete ownership[k];
                     else ownership[k] = v;
                 }
-                let data = { _id: d.id };
-                if (isNewerVersion(game.version, "9.999999"))
-                    data.ownership = ownership;
-                else
-                    data.permission = ownership;
+                let data = { _id: d.id, ownership };
                 return data;
             });